auth api/style

+++ seproject/src/main/java/com/dudungtak/seproject/controller/api/StyleApiController.java
@@ -1,5 +1,7 @@
 package com.dudungtak.seproject.controller.api;
 
+import com.dudungtak.seproject.controller.AuthFilter;
+import com.dudungtak.seproject.enumpackage.AccessType;
 import com.dudungtak.seproject.network.Header;
 import com.dudungtak.seproject.network.request.StyleApiRequest;
 import com.dudungtak.seproject.network.response.StyleApiResponse;
@@ -18,26 +20,41 @@ public class StyleApiController {
 
     @PostMapping("")
     public Header<StyleApiResponse> create(Authentication authentication, @RequestBody Header<StyleApiRequest> request) {
-        return styleApiService.create(authentication, request);
+        if(!AuthFilter.isValidAccess(authentication, AccessType.MANAGER))
+            return Header.ERROR("permission denied");
+
+        return styleApiService.create(request);
     }
 
     @GetMapping("{id}")
-    public Header<StyleApiResponse> read(@PathVariable Long id) {
+    public Header<StyleApiResponse> read(Authentication authentication, @PathVariable Long id) {
+        if(!AuthFilter.isValidAccess(authentication, AccessType.ALL))
+            return Header.ERROR("permission denied");
+
         return styleApiService.read(id);
     }
 
     @GetMapping("")
-    public Header<List<StyleApiResponse>> readAll() {
+    public Header<List<StyleApiResponse>> readAll(Authentication authentication) {
+        if(!AuthFilter.isValidAccess(authentication, AccessType.ALL))
+            return Header.ERROR("permission denied");
+
         return styleApiService.readAll();
     }
 
     @PutMapping("")
-    public Header<StyleApiResponse> update(@RequestBody Header<StyleApiRequest> request) {
+    public Header<StyleApiResponse> update(Authentication authentication, @RequestBody Header<StyleApiRequest> request) {
+        if(!AuthFilter.isValidAccess(authentication, AccessType.MANAGER))
+            return Header.ERROR("permission denied");
+
         return styleApiService.update(request);
     }
 
     @DeleteMapping("{id}")
-    public Header delete(@PathVariable Long id) {
+    public Header delete(Authentication authentication, @PathVariable Long id) {
+        if(!AuthFilter.isValidAccess(authentication, AccessType.MANAGER))
+            return Header.ERROR("permission denied");
+
         return styleApiService.delete(id);
     }
 }
